+++ src/main/java/com/talentotech/final_ecommerce/model/Product.java
@@ -12,7 +12,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "producto_id")
-    private int productoId;
+    private Long productoId;
     private String nombre;
     private String descripcion;
     private double precio;
